Audio Class Separated:its working
